programming/python/misc/pyproject.toml
[project]
name = "misc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "pygments>=2.19.1",
]
